Moved scores functions into their own module.
